refactor(core): a file's modification time is nanoseconds, and says so
The unit was asserted in the adapters and nowhere in the domain. Two
structs declared the stamp as a bare int64, `Unchanged` compared raw
integers, and a walk filling it with `Unix()` seconds compiled clean
and left every file looking changed on every scan.

The stamp is `domain.ModTime`, a type whose doc line states the unit,
built from an instant through `ModTimeOf` and read back as one through
`Time`. It stays an integer: it is a filesystem's stamp to compare and
to order by, not an instant the domain reasons about, and an integer is
what the index column, the schema field and the addresses the window
hands back already hold. A `time.Time` would have left `Fingerprint`
comparable, so every `==` on one would still have built while quietly
depending on a monotonic reading and a location.

`port.SourceText` embeds `domain.Fingerprint` rather than declaring the
file's coordinates a second time.

+++ modules/libs/core/adapter/webui/viewer.go
@@ -158,7 +158,7 @@ func (a *API) GetDocument(
 
 	out := &v1.GetDocumentResponse{
 		Pages:       int32(doc.scan.Pages()),
-		Fingerprint: &v1.Fingerprint{Path: print.path, Size: print.size, Mtime: print.mtime},
+		Fingerprint: &v1.Fingerprint{Path: print.path, Size: print.size, Mtime: int64(print.mtime)},
 	}
 	out.Sheets = make([]*v1.Sheet, out.Pages)
 	for i := range out.Sheets {
